requirements.txt
Flask
SQLAlchemy
psycopg2-binary
bcrypt
PyJWT